refactor(post-sorting): move 'sorting' package from models to services
